safari-600.7-branch/LayoutTests/platform/ios-sim/fast/dom/access-key-iframe-expected.txt

Tests to ensure that accesskey works in iframes and other iframes don't effect current accesskey maps.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Accesskeys should work in an iframe. "iframe1" has both "inputG" and "inputH" (accesskey="c") elements.
FAIL iframe1.contentDocument.getElementById("inputG").focus(); pressAccessKey("c"); targetsOfFocusEvents should be inputG,inputH. Was inputG,inputE.

"inputC" element has an accessKey of "a" and other iframes also have elements with accesskey of "a". An acccesskey should not be overridden by other iframes, so "inputC" should be selected.
PASS document.getElementById("inputB").focus(); pressAccessKey("a"); targetsOfFocusEvents; is ["inputB", "inputC"]

A child iframe, iframe1, has an element with accesskey of "d", which should be ignored.
PASS document.getElementById("inputB").focus(); pressAccessKey("d"); targetsOfFocusEvents is ["inputB"]

An accesskey defined in an ancestor iframe should be ignored. "inputD" has accesskey of "b", which should not be selected from descendant iframes, iframe1 and iframe2.
FAIL iframe1.contentDocument.getElementById("inputG").focus(); pressAccessKey("b"); targetsOfFocusEvents should be inputG. Was inputG,inputD.
FAIL iframe2.contentDocument.getElementById("inputK").focus(); pressAccessKey("b"); targetsOfFocusEvents should be inputK. Was inputK,inputD.
PASS successfullyParsed is true

TEST COMPLETE
